refactor the structure to follow the pattern
