+++ backend/apis/user.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, jsonify
-from api.authentication import authentication_required
-from services.store.dafs.user import get_user, update_user
-from services.notify.notifier import publish_event
+from apis.authentication import authentication_required
+from logics.user import get_user, modify_user
+from logics.event import publish
 from libs.hash.hasher import encrypt
 
 api = Blueprint("user", __name__, url_prefix="/user")
@@ -23,10 +23,10 @@ def retrieve_user(user_id, _):
 def upgister_user(user_id, data):
     if "password" in data and data["password"]:
         data["password"] = encrypt(data["password"])
-    updated_user = update_user(user_id, data)
+    updated_user = modify_user(user_id, data)
     if updated_user is not None:
         del updated_user["password"]
-        if publish_event(user_id, "user.update", updated_user):
+        if publish(user_id, "user.update", updated_user):
             return jsonify({"data": "Successfully update the user information"}), 200
         else:
             return jsonify({"data": "Successfully update the user information but failed to publish sync event"}), 503
