+++ backend/app.py
@@ -1,10 +1,10 @@
 from flask import Flask
-from api import authentication, metadata, user, field, season, measurement, sse
-from services.download import downloader
-from services.store import storage
-from services.mail import mailer
-from services.recommend import recommender
-from services.notify import notifier
+from apis import authentication, metadata, user, field, season, measurement, sse
+from repos.download import downloader
+from repos.store import storage
+from repos.mail import mailer
+from repos.recommend import recommender
+from repos.notify import notifier
 from config import APP
