+++ backend/gunicorn_app.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from api import authentication, metadata, user, field, season, ndvi_raster, measurement, sse
+from apis import authentication, metadata, user, field, season, measurement, sse
 
 
 app = Flask(__name__)
@@ -8,6 +8,5 @@ app.register_blueprint(metadata.api)
 app.register_blueprint(user.api)
 app.register_blueprint(field.api)
 app.register_blueprint(season.api)
-app.register_blueprint(ndvi_raster.api)
 app.register_blueprint(measurement.api)
 app.register_blueprint(sse.api)
