+++ backend/logics/event.py
@@ -1,11 +1,8 @@
-import time
-from flask import Blueprint, Response, request, stream_with_context
-from services.notify.notifier import get_channel, ConnectionError
+from time import sleep
+from repos.notify.notifier import publish_event as publish
+from repos.notify.notifier import get_channel, ConnectionError
 from libs.jwt.token import verify_token
 from config import APP
-
-
-api = Blueprint("sse", __name__, url_prefix="/sse")
 
 
 def stream(auth_token: str):
@@ -16,7 +13,7 @@ def stream(auth_token: str):
             channel = get_channel(user_id)
             while channel is None:
                 channel = get_channel(user_id)
-                time.sleep(2)
+                sleep(2)
             for message in channel.listen():
                 if APP.is_testing or verify_token(auth_token) is not None:
                     if message["type"] == "message":
@@ -30,12 +27,3 @@ def stream(auth_token: str):
     finally:
         if channel is not None:
             channel.close()
-
-
-@api.route("")
-def stream_events():
-    auth_token = request.args.get("auth_token")
-    return Response(stream_with_context(stream(auth_token)),
-                    headers={"Content-Type": "text/event-stream",
-                             "Cache-Control": "no-cache",
-                             "X-Accel-Buffering": "no"})
