+++ backend/repos/store/dafs/subfield.py
@@ -1,7 +1,6 @@
-from services.store.storage import DbCursor, transaction_decorator
-from json import loads as json_parse
-from psycopg2._psycopg import cursor as Cursor
 from typing import Any
+from json import loads as json_parse
+from repos.store.storage import Cursor
 
 
 def __parse_record(record: tuple) -> dict[str, Any] | None:
@@ -20,23 +19,33 @@ def __parse_record(record: tuple) -> dict[str, Any] | None:
     }
 
 
-@transaction_decorator
-def update_subfield_recommended_fertilizer_amount(
-    subfield_id: int,
-    recommended_fertilizer_amount: float,
-    cursor: Cursor | None
-) -> None:
-    cursor.execute(
-        "UPDATE subfield SET recommended_fertilizer_amount = %s WHERE id = %s",
-        (recommended_fertilizer_amount, subfield_id,)
-    )
+def __extract_nonempty(data: dict[str, Any]) -> tuple[list[str], list[Any]]:
+    cols, vals = [], []
+    for col in ["recommended_fertilizer_amount"]:
+        if col in data and data[col] is not None:
+            cols.append(col)
+            vals.append(data[col])
+    return cols, vals
 
 
-@transaction_decorator
+def update_subfield(
+    cursor: Cursor,
+    user_id: int, subfield_id: int,
+    data: dict[str, Any]
+) -> dict[str, Any] | None:
+    cols, vals = __extract_nonempty(data)
+    if len(cols) > 0:
+        update_cols = " = %s, ".join(cols) + " = %s"
+        cursor.execute(
+            f"UPDATE subfield SET {update_cols} WHERE user_id = %s AND id = %s",
+            (*vals, user_id, subfield_id,)
+        )
+
+
 def insert_subfield(
+    cursor: Cursor,
     user_id: int, field_id: int, season_id: str, measurement_id: int,
     region: str, ndvi: float,
-    cursor: Cursor | None = None
 ) -> dict[str, Any] | None:
     cursor.execute(
         """
@@ -59,17 +68,16 @@ def insert_subfield(
     return __parse_record(cursor.fetchone())
 
 
-def list_subfields(user_id: int, field_id: int, season_id: str) -> list[dict[str, Any]] | None:
-    subfields = None
-    db_cursor = DbCursor()
-    with db_cursor as cursor:
-        cursor.execute(
-            """
-            SELECT id, field_id, season_id, measurement_id, ST_AsGeoJSON(region), area, ndvi, recommended_fertilizer_amount
-            FROM subfield
-            WHERE user_id = %s AND field_id = %s AND season_id = %s
-            """,
-            (user_id, field_id, season_id,)
-        )
-        subfields = [__parse_record(record) for record in cursor.fetchall()]
-    return subfields if db_cursor.error is None else None
+def select_subfields(
+    cursor: Cursor,
+    user_id: int, field_id: int, season_id: str
+) -> list[dict[str, Any]] | None:
+    cursor.execute(
+        """
+        SELECT id, field_id, season_id, measurement_id, ST_AsGeoJSON(region), area, ndvi, recommended_fertilizer_amount
+        FROM subfield
+        WHERE user_id = %s AND field_id = %s AND season_id = %s
+        """,
+        (user_id, field_id, season_id,)
+    )
+    return [__parse_record(record) for record in cursor.fetchall()]
